Rename `menu_handle` to `right_click_menu` and `child` to `trigger`
This makes things more in-line with `popover_menu`.
