+++ crates/ui2/src/components.rs
@@ -10,6 +10,7 @@ mod label;
 mod list;
 mod popover;
 mod popover_menu;
+mod right_click_menu;
 mod stack;
 mod tooltip;
 
@@ -28,6 +29,7 @@ pub use label::*;
 pub use list::*;
 pub use popover::*;
 pub use popover_menu::*;
+pub use right_click_menu::*;
 pub use stack::*;
 pub use tooltip::*;
